GeoPressureR: Global positioning by atmospheric pressure.

GeoPressureR is a R package which help researchers determine the position of a bird based on the data retrieved from multi-sensor geolocators.

Details

The main novelty of this package is to match atmospheric pressure measurements to weather reanalysis database (ERA-5) and produce a probability maps of position.

In addition, we use a graphical model to combine pressure, light, acceleration and windspeed data into a trajectory model of the bird.

See also

Author

Maintainer: Raphaël Nussbaumer rafnuss@gmail.com (ORCID)

Authors:

Mathieu Gravey mathieu@mgravey.com (ORCID)