Construct bird morphology

This function return a list with the four morphological information necessary to construct the power curve: mass, wing span, wing aspect ratio and body frontal area.

Usage

flight_bird(
  species_name,
  mass = NA,
  wing_span = NA,
  wing_aspect = NA,
  wing_area = NA,
  body_frontal_area = NA
)

Arguments

species_name

Scientific name of the species

mass

Mass of the bird in kilogram.

wing_span

Wing span in meter.

wing_aspect

Wing aspect ratio (no unit).

wing_area

Wing area in meter square.

body_frontal_area

Body frontal area in meter square.

Value

List containing mass, wing span, wing aspect ratio and body frontal area.

Details

When any of these variables are missing, we query the AVONET database using the scientific name from the Clements Checklist.

Examples

# Using AVONET dataset
flight_bird("Acrocephalus arundinaceus")
#> $species_name
#> [1] "Acrocephalus arundinaceus"
#> 
#> $mass
#> [1] 0.03
#> 
#> $body_frontal_area
#> [1] 0.001498105
#> 
#> $wing_span
#> [1] 0.242256
#> 
#> $wing_aspect
#> [1] 7.114714
#> 
# Using AVONET dataset + custom values
flight_bird("Acrocephalus arundinaceus", wing_aspect = 8)
#> $species_name
#> [1] "Acrocephalus arundinaceus"
#> 
#> $mass
#> [1] 0.03
#> 
#> $body_frontal_area
#> [1] 0.001498105
#> 
#> $wing_span
#> [1] 0.242256
#> 
#> $wing_aspect
#> [1] 8
#> 
# Import your own bird. You will need mass, and at least two of wing_span,
# wing_aspect or wing_area.
flight_bird("custum_bird", mass = 1, wing_span = 1, wing_aspect = 4)
#> $species_name
#> [1] "custum_bird"
#> 
#> $mass
#> [1] 1
#> 
#> $body_frontal_area
#> [1] 0.0129
#> 
#> $wing_span
#> [1] 1
#> 
#> $wing_aspect
#> [1] 4
#>