Convert mismatch map into likelihood map

This function converts the mismatch maps (MSE and mask) into a likelihood map.

Usage

geopressure_map_likelihood(
  tag,
  sd = 1,
  thr_mask = 0.9,
  log_linear_pooling_weight = function(n) log(n)/n
)

Arguments

tag

a GeoPressureR tag object

sd

Standard deviation of the pressure error . numeric of lenght 1 or number of stationary periods.

thr_mask

Threshold of the percentage of data points outside the elevation range to be considered not possible.

log_linear_pooling_weight

Weighting function of the log-linear pooling, taking the number of samples of the stationary periods used and returning the weight of the aggregation. See GeoPressureManual | Probability aggregation for more details.

Details

We convert the map of the mean square error \(MSE\) and altitude mask \(z_{mask}\) computed by geopressure_map_mismatch() into a likelihood map with,

$$L = \exp \left(-w \frac{MSE}{\sigma} \right) \left[z_{mask}>T \right],$$

where \(\sigma\) is the standard deviation of pressure and \(T\) is the mask threshold.

Because the auto-correlation of the timeseries is not accounted for in this equation, we use a log-linear pooling weight of \(w=\log(n)/n\) by default, where \(n\) is the number of samples in the timeserie (i.e., data points used to compute the MSE).

For more background and details on the algorithm, please read Nussbaumer et al. (2023a ) and the GeoPressureManual | Probability aggregation .

References

Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, and Felix Liechti. 2023. Global Positioning with Animal‐borne Pressure Sensors. Methods in Ecology and Evolution, 14, 1118–1129 https://doi.org/10.1111/2041-210X.14043.

See also