Request and download mismatch maps from pressure data

This function returns, for each stationary period, two maps of mismatch between the pressure measured by the geolocator and the ERA5 pressure database.

Usage

geopressure_map_mismatch(
  tag,
  max_sample = 250,
  margin = 30,
  timeout = 60 * 5,
  workers = "auto",
  compute_known = FALSE,
  debug = FALSE,
  quiet = FALSE
)

Arguments

tag

a GeoPressureR tag object

max_sample

The computation of the maps is only performed on max_sample datapoints of pressure to reduce computational time. The samples are randomly (uniformly) selected on the timeseries.

margin

The margin is used in the mask map to accept measurement errors, small-scale topography, and vertical movements of the bird (unit in meters, 1hPa~10m).

timeout

Duration before the code is interrupted both for the request on GeoPressureAPI and GEE (in seconds, see httr::timeout()).

workers

Number of parallel requests on GEE. Integer between 1 and 99. "auto" adjust the number of workers to the number of stap_elev to query.

compute_known

logical defining if the map(s) for known stationary period should be estimated based on twilight or hard defined by the known location stap$known_l**

debug

logical to display additional information to debug a request

quiet

logical to hide messages about the progress

Details

These maps are generated on Google Earth Engine via the map entry point of the GeoPressure API. The computation performed by this function consists of the following:

Send a request to generate the Google Earth Engine (GEE) URL of the code producing the maps for each stationary period separately.

Download and read these geotiff maps as matrix. The computation on GEE only happens in this second step when fetching the URL.

The maps of each stationary period are returned in two layers:

The Mean Square Error (MSE) between the data logger pressure timeseries and the reanalysis. The mean error is removed because we assume no specific altitude of the geolocator, thus allowing an altitudinal shift of the pressure timeseries.

The mask of the proportion of pressure measurements corresponding to altitude values found within the min and max ground elevation at each location. The altitude value of the geolocator pressure timeseries is computed with the barometric formula accounting for the temporal variation of pressure (surface-pressure) and temperature (2m-temperature) based on ERA5 data. The min and max ground elevation of each pixel is computed from SRTM-90.

For each stationary period, the pressure measurements are smoothed and downscaled to a 1-hour resolution in order to match ERA-5 resolution.

It is possible to indicate different elevation levels when the bird was spending time at locations with different elevations within a general area (~10km), and thus within the same stationary period. This can be done by using tag$label="elev_n"for all measurements of the same elevation level n.

For more background and details on this algorithm, please refer to Nussbaumer et al. (2023a ) and the GeoPressure API documentation .

References

Nussbaumer, Raphaël, Mathieu Gravey, Martins Briedis, and Felix Liechti. 2023. Global Positioning with Animal‐borne Pressure Sensors. Methods in Ecology and Evolution, 14, 1118–1129 https://doi.org/10.1111/2041-210X.14043.

See also