Start the GeoPressureViz shiny app

GeoPressureViz is a shiny app designed to help you visualize the overall trajectory of the bird as well as each step-by-step move. This app is particularly useful to check the correspondence between pressure map, light map and flight distance. You can edit the path and query pressure time series for individual stationary period to test manual what seems the optimal path.

Usage

geopressureviz(x, path = NULL, marginal = NULL, launch_browser = TRUE)

Arguments

x

a GeoPressureR tag object, a .Rdata file or the unique identifier id with a .Rdata file located in "./data/interim/{id}.RData".

path

a GeoPressureR path or pressurepath data.frame.

marginal

map of the marginal probability computed with graph_marginal(). Overwrite the path or pressurepath contained in the .Rdata file.

launch_browser

If true (by default), the app runs in your browser, otherwise it runs on Rstudio.

Value

The updated path visualized in the app. Can also be retrieved with .GlobalEnv$path_geopressureviz

Details

GeoPressureViz can be started based on a .Rdata file containing at least tag, but also optionally marginal and/or path (path_most_likely is also accepted).

You can retrieved the edited path from the global environment variable path_geopressureviz.

Learn more about GeoPressureViz in the GeoPressureManual or with this demo of the Great Reed Warbler (18LX).

See also