Compute transition probabilities of a graph

Use the movement model (see graph_set_movement()) to convert ground speed gs (or airspeed as if available) into the transition probability of the edges of the graph.

Usage

graph_transition(graph)

Arguments

graph

Graph constructed with graph_create() and with a movement (see graph_set_movement()).

Value

Vector of transition probability for each edge.

Details

The vector return correspond to the elements of the transition matrices \(T_k \forall k \in [1,n]\) extracted for all edges considered in the graph. Each of these values thus corresponds to the probability \(P(X_k \mid X_{k-1})\), where \(X_k\) is the random variable of the position of the bird at time \(k\).

To create a generic function, we define speed2prob which converts the speed of an edge into the transition probability.

See also