Plot acceleration data of a tag

This function display a plot of acceleration time series recorded by a tag

Usage

plot_tag_acceleration(
  tag,
  plot_plotly = TRUE,
  label_auto = TRUE,
  min_duration = 30
)

Arguments

tag

a GeoPressureR tag object

plot_plotly

logical to use plotly

label_auto

logical to compute and plot the flight label using tag_label_auto(). Only if labels are not already present on tag$acceleration$label

min_duration

Minimal duration (in minutes) to consider a high activity as migratory flight.

Value

a plot or ggplotly object.

See also

Examples

setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE)

plot_tag_acceleration(tag)