Plot temperature data of a tag

This function display a plot of temperature time series recorded by a tag

Usage

plot_tag_temperature(
  tag,
  variable = "external",
  plot_plotly = TRUE,
  label_auto = TRUE,
  min_duration = 30
)

Arguments

tag

a GeoPressureR tag object

variable

temperature variable to plot "external" or "internal"

plot_plotly

logical to use plotly

label_auto

logical to compute and plot the flight label using tag_label_auto(). Only if labels are not already present on tag$temperature$label

min_duration

Minimal duration (in minutes) to consider a high activity as migratory flight.

Value

a plot or ggplotly object.

See also

Examples

withr::with_dir(system.file("extdata", package = "GeoPressureR"), {
  tag <- tag_create("18LX", quiet = TRUE)
})

plot_tag_temperature(tag)