Edit classification of activity and pressure

This function perform three steps: (1) write the csv file of the automatically labeled activity and pressure with trainset_write, (2) open trainset in your broweser (https://trainset.geocene.com/) so that you can edit the labels and (3) read the exported csv file from trainset with trainset_read.

Usage

trainset_edit(
  pam,
  pathname,
  filename = paste0(pam$id, "_act_pres-labeled.csv")
)

Arguments

pam

pam logger dataset list

pathname

Path to the folder where the labeled files should be saved

filename

Name for the file.

Value

pam logger dataset list updated with the new label named class(pam$pressure$class and pam$acceleration$class)