Create bird flight traits

This function returns a list with the four morphological information necessary to construct the power curve: mass, wing span, wing aspect ratio, and body frontal area.

When any of these variables are missing, we query the AVONET database using the scientific name from the Clements Checklist.

Usage

bird_create(
  species_name,
  mass = NULL,
  wing_span = NULL,
  wing_aspect = NULL,
  wing_area = NULL,
  body_frontal_area = NULL
)

Arguments

species_name

Scientific name of the species

mass

Mass of the bird in kilograms.

wing_span

Wing span in meters.

wing_aspect

Wing aspect ratio (no unit).

wing_area

Wing area in meter square.

body_frontal_area

Body frontal area in meter square.

Value

List containing mass, wing span, wing aspect ratio, and body frontal area.

See also

Examples

# Using AVONET dataset
bird_create("Acrocephalus arundinaceus")
#> 
#> ── GeoPressureR `bird` object ──────────────────────────────────────────────────
#> • Species: Acrocephalus arundinaceus
#> • Mass: 0.03 (kg).
#> • Body frontal area: 0 (m^2).
#> • Wing span: 0.2 (m).
#> • Wing aspect: 7.1 (-).

# Using AVONET dataset + custom values
bird_create("Acrocephalus arundinaceus", wing_aspect = 8)
#> 
#> ── GeoPressureR `bird` object ──────────────────────────────────────────────────
#> • Species: Acrocephalus arundinaceus
#> • Mass: 0.03 (kg).
#> • Body frontal area: 0 (m^2).
#> • Wing span: 0.2 (m).
#> • Wing aspect: 8 (-).

# Import your own bird. You will need mass, and at least two of wing_span,
# wing_aspect or wing_area.
bird_create("custum_bird", mass = 1, wing_span = 1, wing_aspect = 4)
#> 
#> ── GeoPressureR `bird` object ──────────────────────────────────────────────────
#> • Species: custum_bird
#> • Mass: 1 (kg).
#> • Body frontal area: 0 (m^2).
#> • Wing span: 1 (m).
#> • Wing aspect: 4 (-).