Return the most likely path from a probability map

Find the location of the highest value in the map and return a path data.frame containing the latitude and longitude. interp can be used to interpolate unrealistic position from short stationary period based on the position of the longer ones. The interpolation assumes that the first and last stationary period can be safely estimated from the probability map.

Usage

geopressure_map2path(map, interp = 0, format = "lonlat")

Arguments

map

list of raster containing probability map of each stationary period. The metadata of map needs to include the start and end time of the stationary period .

interp

(in days) The position of the stationary period shorter than interp will be replace by a linear average from other position.

format

one of "lonlat", "ind", "arr.ind"). return the path in lon-lat or indices

Value

a data.frame of the position containing latitude (lat), longitude (lon) and the stationary period id (sta_id) as column. Optionally, if indexes were requested, it will be return. You will need to use which.max(as.matrix(raster)) and not which.max(raster) to get the correct location

Examples

load(system.file("extdata", "18LX_pressure_prob.rda", package = "GeoPressureR"))
path_all <- geopressure_map2path(pressure_prob)
path_interp <- geopressure_map2path(pressure_prob, interp = 2)
sta_duration <- unlist(lapply(pressure_prob, function(x) {
  as.numeric(difftime(raster::metadata(x)$temporal_extent[2],
    raster::metadata(x)$temporal_extent[1],
    units = "days"
  ))
}))
m <- leaflet::leaflet(width = "100%")
m <- leaflet::addProviderTiles(m, leaflet::providers$Stamen.TerrainBackground)
m <- leaflet.extras::addFullscreenControl(m)
m <- leaflet::addPolylines(m,
  lng = path_all$lon, lat = path_all$lat, opacity = 1,
  color = "#a6cee3", weight = 3
)
m <- leaflet::addCircles(m,
  lng = path_all$lon, lat = path_all$lat, opacity = 1,
  color = "#1f78b4", weight = sta_duration^(0.3) * 10
)
m <- leaflet::addPolylines(m,
  lng = path_interp$lon, lat = path_interp$lat, opacity = 1,
  color = "#b2df8a", weight = 3
)
m <- leaflet::addCircles(m,
  lng = path_interp$lon, lat = path_interp$lat, opacity = 1,
  color = "#33a02c", weight = sta_duration^(0.3) * 10
)
m
#> Error in x$width %||% settings$fig.width * settings$dpi: non-numeric argument to binary operator