Start the GeoPressureViz shiny app

GeoPressureViz is a shiny app designed to helps you visualize the overall trajectory of the bird as well as each step-by-step move. Learn more about GeoPressureViz in the GeoPressureManual | GeoPressureViz or with a demo of the Great Reed Warbler (18LX).

Usage

geopressureviz(
  pam,
  static_prob,
  pressure_prob = NA,
  light_prob = NA,
  static_prob_marginal = NA,
  pressure_prob_thr = NA,
  pressure_prob_mismatch = NA,
  pressure_timeserie = NA,
  lauch_browser = TRUE
)

Arguments

pam

PAM logger dataset list with pam$sta computed. See pam_read() and pam_sta().

static_prob

List of raster containing probability map of each stationary period. The metadata of static_prob needs to include the flight information to the next stationary period in the metadata flight. See GeoPressureManual | Static map .

pressure_prob

List of raster containing probability map of each stationary period according to pressure data. See GeoPressureManual | Pressure map .

light_prob

List of raster containing probability map of each stationary period according to light data. See GeoPressureManual | Light map .

static_prob_marginal

List of raster containing probability map of each stationary period according to the graph output. See GeoPressureManual | Basic graph .

pressure_prob_thr

List of raster containing probability map of each stationary period according to the threshold of pressure data. See GeoPressureManual | Pressure map .

pressure_prob_mismatch

List of raster containing probability map of each stationary period according to mismatch of pressure data. See GeoPressureManual | Pressure map .

pressure_timeserie

List of data.frame containing at least sta_id, date and pressure0

lauch_browser

If true (default), the app run in your browser, otherwise on Rstudio

Value

The path modified in the app.

See also

Examples

if (FALSE) {
load("data/1_pressure/18LX_pressure_prob.Rdata")
load("data/2_light/18LX_light_prob.Rdata")
load("data/3_static/18LX_static_prob.Rdata")
geopressureviz(
  pam = pam,
  static_prob = static_prob,
  pressure_prob = pressure_prob,
  light_prob = light_prob,
  pressure_timeserie = static_timeserie
)
}