Create a param list

Create the list of parameter for GeoPressureR tag and graph objects.

param list are mostly used to archived the actual value of parameters used to create a tag and/or a graph, thus allowing for examination of parameters post-creation. This function should therefore not be used to set/define parameters ahead of computation. In reality, there are very few external case of use for this function.

Usage

param_create(id, default = FALSE, ...)

Arguments

id

Unique identifier of a tag.

default

logical to initiate param with default value of the package.

...

arguments passed to other methods.

Value

A GeoPressureR param list

See also

Examples

param <- param_create("18LX", extent = c(0, 0, 1, 1))
print(param)
#> 
#> ── GeoPressureR `param` object for id: 18LX ────────────────────────────────────
#> Note: All green texts are fields of `param` (i.e., `param$field`).
#> • GeoPressureR_version: 3.1.7
#> 
#> ── Sensors data `tag_create()` 
#> • manufacturer:
#> • crop_start:
#> • crop_end:
#> • directory:
#> • pressure_file:
#> • light_file:
#> • acceleration_file:
#> 
#> ── tag label `tag_label()` 
#> • label_file:
#> 
#> ── Stationary period definition `tag_set_map()` 
#> • known:
#> NULL
#> • include_stap_id:
#> • include_min_duration:
#> 
#> ── Geographical parameters `geo_expend()` 
#> • extent: 0, 0, 1, and 1
#> • scale:
#> 
#> ── Geopressure `geopressure_map()` 
#> • max_sample:
#> • margin:
#> • sd:
#> • thr_mask:
#> • log_linear_pooling_weight:
#> 
#> ── Twilight & Geolight `twilight_create()` 
#> • twl_thr:
#> • twl_offset:
#> • twilight_file:
#> • twl_calib_adjust:
#> • twl_llp:
#> 
#> ── Graph `graph_create()` 
#> • thr_likelihood:
#> • thr_gs:
#> 
#> ── Movement model & wind `graph_add_wind()` `graph_movement()` 
#> • thr_as:
#> • wind_file:
#> • type:
#> • method:
#> • shape:
#> • scale:
#> • location:
#> • bird:
#> • power2prob:
#> • low_speed_fix: