Create a pressurepath

A pressurepath is a data.frame representing a timeseries of ERA5 reanalysis pressure along a path (position over time). It can be thought as the equivalent of the hypothetical measurement of a pressure sensor equiped on a bird moving along a specific path.

pressurepath are useful for labeling tag data by allowing a direct comparison of the actual pressure measured by the sensor to the one equivalent to the best estimate of the path.

You can include previous and/or next flight period in each query, for instance to estimate flight altitude.

The function essentially splits the pressure of the geolocator into stationary periods and downloads the ERA5 timeseries of pressure for the corresponding location of the path by calling geopressure_timeseries() at each stationary period. The computation is optimized by paralleling the call of each stationary period.

Usage

pressurepath_create(
  tag,
  path = tag2path(tag),
  include_flight = FALSE,
  workers = "auto",
  preprocess = FALSE,
  quiet = FALSE
)

Arguments

tag

a GeoPressureR tag object.

path

a GeoPressureR path data.frame.

include_flight

extend request to also query the pressure and altitude during the previous and/or next flight. Flights are defined by a stap_id = 0. Accept logical or vector of -1 ( previous flight), 0 (stationary) and/or 1 (next flight). (e.g. include_flight = c(-1, 1) will only search for the flight before and after but not the stationary period). Note that next and previous flights are defined by the +/1 of the stap_id value (and not the previous/next stap_id value).

workers

number of parallel requests on GEE. Between 1 and 99. "auto" adjust the number of workers to the number of stap_elev to query.

preprocess

logical to use geopressure_map_preprocess.

quiet

logical to hide messages about the progress

Value

A GeoPressureR pressurepath data.frame with columns:

date same as pressure$date

pressure_tag: same as pressure$value

label same as pressure$label

stap_id same as pressure$stap_id (i.e., including 0)

pressure_era5 pressure retrieved from ERA5.

altitude Altitude computed

lat same as path$lat

lon same as path$lon

pressure_era5_norm pressure retrieved from ERA5 normalized to the average of pressure_tag over the stationary period.

stap_ref stap_id grouping of the requests (same as stap_id except for flights where stap_id = 0 and stap_ref is to the stationary period of the position used from the path).

See also

Examples

setwd(system.file("extdata", package = "GeoPressureR"))
tag <- tag_create("18LX", quiet = TRUE) |> tag_label(quiet = TRUE)

path <- data.frame(
  stap_id = tag$stap$stap_id,
  lat = c(48.5, 32.5, 30.5, 49.5, 41.6),
  lon = c(17.5, 13.5, 16.5, 21.5, 12.7)
)

pressurepath <- pressurepath_create(tag, path = path, quiet = TRUE)

str(pressurepath)
#> 'data.frame':	624 obs. of  10 variables:
#>  $ date              : POSIXct, format: "2017-07-27 00:00:00" "2017-07-27 00:30:00" ...
#>  $ pressure_tag      : int  989 989 990 990 989 989 990 990 991 990 ...
#>  $ label             : chr  "" "" "" "" ...
#>  $ stap_id           : num  1 1 1 1 1 1 1 1 1 1 ...
#>  $ pressure_era5     : num  978 978 978 978 978 ...
#>  $ altitude          : num  157 157 148 148 158 ...
#>  $ lat               : num  48.5 48.5 48.5 48.5 48.5 48.5 48.5 48.5 48.5 48.5 ...
#>  $ lon               : num  17.5 17.5 17.5 17.5 17.5 17.5 17.5 17.5 17.5 17.5 ...
#>  $ pressure_era5_norm: num  989 989 989 989 989 ...
#>  $ stap_ref          : num  1 1 1 1 1 1 1 1 1 1 ...
#>  - attr(*, "id")= chr "18LX"
#>  - attr(*, "preprocess")= logi FALSE
#>  - attr(*, "include_flight")= num 0

plot_pressurepath(pressurepath)

pressurepath <- pressurepath_create(
  tag,
  path[c(2, 3, 4), ],
  include_flight = TRUE,
  quiet = TRUE
)

plot_pressurepath(pressurepath)