Download flight data

This function download data associated to each flight from the ERA5 hourly pressure levels with the Climate Data Store (CDS) and through the ecmwfr R package.

Any variable available from the ERA5 pressure level can be downloaded.

The flights are determined from the stationary periods classified tag$stap. It request a single file for each flight using the exact time (hourly basis) and pressure (altitude). To make the download more efficient, wf_request_batch() is used to download all files at the same time (up to 20 requests in parallel).

To be able to download data from the Climate Data Store (CDS), you will need to create an account on https://cds.climate.copernicus.eu. Once created, you can retrieve your API key on https://cds.climate.copernicus.eu/user/ and save them in your environment file (i.e., .Renviron). You can open this file with usethis::edit_r_environ() and add: cds_user = "{UID}" cds_key = "{API Key}"

Usage

tag_download_wind(
  tag,
  extent = tag$param$extent,
  include_stap_id = NULL,
  variable = c("u_component_of_wind", "v_component_of_wind"),
  cds_key = Sys.getenv("cds_key"),
  cds_user = Sys.getenv("cds_user"),
  file = function(stap_id)
    glue::glue("./data/wind/{tag$param$id}/{tag$param$id}_{stap_id}.nc"),
  overwrite = FALSE
)

Arguments

tag

a GeoPressureR tag object.

extent

geographical extent of the map on which the likelihood will be computed. Vector of length 4 c(xmin, xmax, ymin, ymax) or c(W, E, S, N).

include_stap_id

stationary period identifiers of the start of the flight to download. Default is to download all flights.

variable

list of variables to download from the ERA5 pressure level: "u_component_of_wind", "v_component_of_wind", "temperature", "fraction_of_cloud_cover", "relative_humidity", , "vertical_velocity", "specific_cloud_ice_water_content", "specific_cloud_liquid_water_content", "specific_humidity", "specific_rain_water_content", "specific_snow_water_content", "divergence", "geopotential", "ozone_mass_mixing_ratio", "potential_vorticity", 'vorticity".

cds_key

ECMWF API Key available from your user page . See wf_set_key().

cds_user

ECMWF user name (UID value) available from your user page . See wf_set_key().

file

absolute or relative path of the ERA5 wind data file to be downloaded. Function taking as single argument the stationary period identifier.

overwrite

logical. If TRUE, file is overwritten.

Value

the path of the downloaded (requested file) or the an R6 object with download/transfer information

See also